+++ Euler_46.cpp
@@ -4,13 +4,13 @@ llui Euler::GoldbachsOtherConjecture()
 {
     std::vector<int> primes = EulerUtility::getPrimesUnderCeilingIndexed(1000000);
 
-    for (int i = 33; i < primes.size(); ++i)
+    for (uint64_t i = 33; i < primes.size(); ++i)
     {
         if ((i & 1) && primes[i] == -1)
         {
             bool conjecture_holds = false;
 
-            for (int j = 0; j < i; ++j)
+            for (uint64_t j = 0; j < i; ++j)
             {
                 if (primes[j] != -1)
                 {
