+++ Euler_49.cpp
@@ -14,7 +14,7 @@ std::string Euler::PrimePermutations()
             bool isUniquePermutation = true;
             std::string s = std::to_string(p);
 
-            for (int i = 0; i < primePermutations.size(); ++i)
+            for (uint64_t i = 0; i < primePermutations.size(); ++i)
             {
                 if (std::is_permutation(primePermutations[i][0].begin(), primePermutations[i][0].end(), s.begin()))
                 {
@@ -32,9 +32,9 @@ std::string Euler::PrimePermutations()
 
     for (std::vector<std::string> pp : primePermutations)
         if (pp.size() > 2)
-            for (int i = 0; i < pp.size() - 2; ++i)
-                for (int j = i + 1; j < pp.size() - 1; ++j)
-                    for (int k = j + 1; k < pp.size(); ++k)
+            for (uint64_t i = 0; i < pp.size() - 2; ++i)
+                for (uint64_t j = i + 1; j < pp.size() - 1; ++j)
+                    for (uint64_t k = j + 1; k < pp.size(); ++k)
                         if ((atoi(pp[j].c_str()) * 2) - atoi(pp[i].c_str()) == atoi(pp[k].c_str()))
                             return pp[i] + pp[j] + pp[k];
